beauty-bot/requirements.txt
aiogram==3.13.1
google-genai==1.14.0
aiosqlite==0.20.0
python-dotenv==1.0.1
pillow==11.1.0
